first working commit
provides a baseline Django app for a bootstrap calendar. Allows for the
creation of new events without rerendering of the page and a relatively
responsive interface.
things to be improved: optimization and debugging might be needed for
various features and JavaScript needs to be cleaned, separated from html
and minimized. addition of update and deletion will be included in the
next commit.